refactor(likes): share request context setup
